+++ doomed_patient/doomed_patient/CellPassage.cpp
@@ -4,9 +4,8 @@
 #include "PatientGame.h"
 
 
-CellPassage::CellPassage(Directions::Direction direction, std::shared_ptr<LevelCell> cell, PatientGame* game, bool isDoor)
-	: CellEdge(direction, cell),
-	isDoor(isDoor)
+CellPassage::CellPassage(Directions::Direction direction, std::shared_ptr<LevelCell> cell, PatientGame* game)
+	: CellEdge(direction, cell)
 {
 	initialiseSprite(game);
 	initialisePosition();
@@ -18,20 +17,5 @@ CellPassage::~CellPassage()
 
 void CellPassage::initialiseSprite(PatientGame* game)
 {
-	if (isDoor)
-	{
-		if (direction == Directions::Direction::NORTH || direction == Directions::Direction::SOUTH)
-		{
-			sprite = game->getHorizontalDoorSprite();
-		}
-
-		else if (direction == Directions::Direction::EAST || direction == Directions::Direction::WEST)
-		{
-			sprite = game->getVerticalDoorSprite();
-		}
-	}
-	else
-	{
 		sprite = nullptr;
-	}
 }